% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_mix_known.R
\name{em_mix_known}
\alias{em_mix_known}
\title{EM Implementation for ancestry mixtures}
\usage{
em_mix_known(x, dat, pnames, pi_init, Ntot, threshold = 0.01,
  MAF_thresh = 0.05, path = NA)
}
\arguments{
\item{x}{N by 3 data.frame for N markers and 3 genotype columns (e.g. "hom", "homref" and "het") giving allele frequencies of those observed at genetic marker.}

\item{dat}{data.frame object for frequencies by each population. This is the reference dataset that will be used.}

\item{pnames}{matrix of population names in column 1 that contains K rows (one for each ancestry considered), followed by the column names pertaining to the allele frequency(ies) for that population as they appear in dat.}

\item{pi_init}{numeric vector of length K providing initial mixture parameters.}

\item{Ntot}{total number of population considered for x.}

\item{threshold}{numeric vector of length 1 indicating when to terminate the loop. The loop will terminate when the sum of the absolute differences between the previous and current estimates for the population mixture is less than \code{threshold}.}

\item{MAF_thresh}{numeric vector of length 1 with value between 0 and 0.5 indicating when to exclude a marker due to rarity (currently not implemented).}

\item{path}{gives a file path name for the directory where tables of outputs from the iterations should be written.}
}
\value{
EMmix object recording the iteration results from the function.
}
\description{
Implements the EM algorithm to estimate the mixture parameters of ancestries in an observed dataset using a reference dataset of some sort.
}
