% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_object.R
\docType{class}
\name{EMmix-class}
\alias{EMmix-class}
\alias{EMmix}
\title{EMmix Object}
\arguments{
\item{pnames}{a matrix of ancestry names in the first column and the corresponding column names for each genotype.}

\item{x}{the observed data.}

\item{dat}{the reference data that is used to determine the mixture proportions in x.}

\item{pi_init}{the initiating mixture proportions.}

\item{med}{the median values of the estimated mixture proportions of each ancestry from each EM iteration.}

\item{mu}{the mean values of the estimated mixture proportions of each ancestry from each EM iteration.}

\item{x90}{the upper 90th quantile values of the estimated mixture proportions of each ancestry from each EM iteration.}

\item{geneMixes}{the final estimates for the genetic inheritances. Each row indicates a different gene, and each column indicates an ancestry.}
}
\description{
Records information from the EM iterations when running em_mix_known
}
