% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_object.R
\docType{methods}
\name{plot,EMmix,missing-method}
\alias{plot,EMmix,missing-method}
\title{Plot EM Object}
\usage{
\S4method{plot}{EMmix,missing}(x, y)
}
\arguments{
\item{x}{an EMmix object}
}
\description{
Plots the iteration results to display the rate of convergence of the algorithm to the final results
}
