% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_object.R
\docType{methods}
\name{show,EMmix-method}
\alias{show,EMmix-method}
\title{Show EMmix Object}
\usage{
\S4method{show}{EMmix}(object)
}
\arguments{
\item{object}{an EMmix object}
}
\description{
Show function for an EM mix object
}
